+++ rust/src/instr_id_enum.rs
@@ -154,6 +154,7 @@ pub enum InstrId {
     cpu_tlbwi,
     cpu_tlbwr,
     cpu_tlbp,
+    cpu_rfe,
     cpu_eret,
     cpu_mfc1,
     cpu_dmfc1,