+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Implement the `rfe` ("Restore From Exception") instruction.
+
 ## [1.11.1] - 2024-07-12
 
 ### Added